django/requirements.txt
django==4.1
django-braces
django-reset-migrations
pymupdf
toml
model_bakery # for tests
beautifulsoup4 # HTML parsing
django-session-timeout
django-htmx
tabulate # display table in command line
plom
tk
reportlab
django-huey
django-polymorphic
whitenoise
